fix: move gantt story to own block
